Include js tag in html
NOTE: it's extremely unsafe having the API key just floating aroung like
this in the repo.  Would suggest using something like Figaro to mask
that, but given the ease of use for you, O Interviewer, I figured it'd
be easier to hardcode it here.

The key is linked to a throw-away account anyway